Fix tooltip rendering
