+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Removed
 
 ### Fixed
+- Fix a bug where tooltips no longer render on hover.
 
 ### Security
